+++ src/factions/sons_of_behemat/index.ts
@@ -1,6 +1,8 @@
 import { Faction } from 'factions/factionClass'
+import { pickEffects } from 'factions/metatagger'
 import { DESTRUCTION } from 'meta/alliances'
 import { SONS_OF_BEHEMAT } from 'meta/factions'
+import battle_traits from './battle_traits'
 import rule_sources from './rule_sources'
 import SubFactions from './subfactions'
 
@@ -9,5 +11,6 @@ export const SonsOfBehematFaction = new Faction(
   DESTRUCTION,
   SubFactions,
   'Tribes',
-  rule_sources.BATTLETOME_SONS_OF_BEHEMAT
+  rule_sources.BATTLETOME_SONS_OF_BEHEMAT,
+  pickEffects(battle_traits, [SONS_OF_BEHEMAT])
 )
