+++ src/factions/sons_of_behemat/subfactions.ts
@@ -1,7 +1,5 @@
-import { pickEffects } from 'factions/metatagger'
 import { SONS_OF_BEHEMAT } from 'meta/factions'
 import Artifacts from './artifacts'
-import battle_traits from './battle_traits'
 import command_abilities from './command_abilities'
 import CommandTraits from './command_traits'
 import Flavors from './flavors'
@@ -9,7 +7,7 @@ import Units from './units'
 
 const subFactions = {
   [SONS_OF_BEHEMAT]: {
-    effects: pickEffects(battle_traits, ['Sons of Behemat']),
+    effects: [],
 
     available: {
       artifacts: [Artifacts],
