+++ src/factions/tzeentch/index.ts
@@ -1,6 +1,8 @@
 import { Faction } from 'factions/factionClass'
+import { pickEffects } from 'factions/metatagger'
 import { CHAOS } from 'meta/alliances'
 import { TZEENTCH } from 'meta/factions'
+import battle_traits from './battle_traits'
 import rule_sources from './rule_sources'
 import SubFactions from './subfactions'
 
@@ -9,5 +11,6 @@ export const TzeentchFaction = new Faction(
   CHAOS,
   SubFactions,
   'Change Covens',
-  rule_sources.BATTLETOME_TZEENTCH
+  rule_sources.BATTLETOME_TZEENTCH,
+  pickEffects(battle_traits, ['The Flow of Change'])
 )
