+++ src/factions/wanderers/subfactions.ts
@@ -1,8 +1,6 @@
-import { pickEffects } from 'factions/metatagger'
 import { WANDERERS } from 'meta/factions'
 import Artifacts from './artifacts'
 import Battalions from './battalions'
-import BattleTraits from './battle_traits'
 import CommandAbilities from './command_abilities'
 import CommandTraits from './command_traits'
 import Spells from './spells'
@@ -10,7 +8,7 @@ import Units from './units'
 
 const subFactions = {
   [WANDERERS]: {
-    effects: pickEffects(BattleTraits, ['Wanderers']),
+    effects: [],
 
     available: {
       artifacts: [Artifacts],
